+++ cmsdev/internal/test/ims/api.go
@@ -31,65 +31,10 @@ package ims
 import (
 	"encoding/json"
 	"net/http"
+
 	"stash.us.cray.com/SCMS/cms-tools/cmsdev/internal/lib/common"
 	"stash.us.cray.com/SCMS/cms-tools/cmsdev/internal/lib/test"
 )
-
-// Return specific image record in IMS via API
-func getIMSImageRecordAPI(imageId string) (imageRecord IMSImageRecord, ok bool) {
-	var baseurl string = common.BASEURL
-	ok = false
-
-	common.Infof("Getting image record %s in IMS via API", imageId)
-	params := test.GetAccessTokenParams()
-	if params == nil {
-		return
-	}
-	url := baseurl + endpoints["ims"]["images"].Url + "/" + imageId
-	resp, err := test.RestfulVerifyStatus("GET", url, *params, http.StatusOK)
-	if err != nil {
-		common.Error(err)
-		return
-	}
-
-	// Extract image record from response
-	common.Infof("Decoding JSON in response body")
-	if err := json.Unmarshal(resp.Body(), &imageRecord); err != nil {
-		common.Error(err)
-		return
-	}
-	ok = true
-
-	return
-}
-
-// Return a list of all image records in IMS via API
-func getIMSImageRecordsAPI() (recordList []IMSImageRecord, ok bool) {
-	var baseurl string = common.BASEURL
-	ok = false
-
-	common.Infof("Getting list of all image records in IMS via API")
-	params := test.GetAccessTokenParams()
-	if params == nil {
-		return
-	}
-	url := baseurl + endpoints["ims"]["images"].Url
-	resp, err := test.RestfulVerifyStatus("GET", url, *params, http.StatusOK)
-	if err != nil {
-		common.Error(err)
-		return
-	}
-
-	// Extract list of image records from response
-	common.Infof("Decoding JSON in response body")
-	if err := json.Unmarshal(resp.Body(), &recordList); err != nil {
-		common.Error(err)
-		return
-	}
-	ok = true
-
-	return
-}
 
 // Return specific job record in IMS via API
 func getIMSJobRecordAPI(jobId string) (jobRecord IMSJobRecord, ok bool) {
@@ -137,62 +82,6 @@ func getIMSJobRecordsAPI() (recordList []IMSJobRecord, ok bool) {
 	}
 
 	// Extract list of job records from response
-	common.Infof("Decoding JSON in response body")
-	if err := json.Unmarshal(resp.Body(), &recordList); err != nil {
-		common.Error(err)
-		return
-	}
-	ok = true
-
-	return
-}
-
-// Return specific public key record in IMS via API
-func getIMSPublicKeyRecordAPI(pkeyId string) (pkeyRecord IMSPublicKeyRecord, ok bool) {
-	var baseurl string = common.BASEURL
-	ok = false
-
-	common.Infof("Getting public key record %s in IMS via API", pkeyId)
-	params := test.GetAccessTokenParams()
-	if params == nil {
-		return
-	}
-	url := baseurl + endpoints["ims"]["public_keys"].Url + "/" + pkeyId
-	resp, err := test.RestfulVerifyStatus("GET", url, *params, http.StatusOK)
-	if err != nil {
-		common.Error(err)
-		return
-	}
-
-	// Extract public key record from response
-	common.Infof("Decoding JSON in response body")
-	if err := json.Unmarshal(resp.Body(), &pkeyRecord); err != nil {
-		common.Error(err)
-		return
-	}
-	ok = true
-
-	return
-}
-
-// Return a list of all public key records in IMS via API
-func getIMSPublicKeyRecordsAPI() (recordList []IMSPublicKeyRecord, ok bool) {
-	var baseurl string = common.BASEURL
-	ok = false
-
-	common.Infof("Getting list of all public key records in IMS via API")
-	params := test.GetAccessTokenParams()
-	if params == nil {
-		return
-	}
-	url := baseurl + endpoints["ims"]["public_keys"].Url
-	resp, err := test.RestfulVerifyStatus("GET", url, *params, http.StatusOK)
-	if err != nil {
-		common.Error(err)
-		return
-	}
-
-	// Extract list of public key records from response
 	common.Infof("Decoding JSON in response body")
 	if err := json.Unmarshal(resp.Body(), &recordList); err != nil {
 		common.Error(err)